[refactor]: 코드 분리 및 initialState 변경
api 도메인 분리

새로고침 시 리덕스 초기화 문제
- 로컬스토리지에 empId, compId를 저장하고,
  initialState를 로컬스토리지에 있는 값을
  가져오는 방식 추가
