+++ src/api/login.js
@@ -11,3 +11,14 @@ export const loginApi = async ({ data }) => {
     console.error('로그인 실패',)
   }
 }
+
+export function changeEmpApi(empId) {
+  if(empId !== undefined && empId) {
+      return axiosInstance.post(
+          `/user/re-login`,
+          {
+            empId,
+          }
+      )
+  }
+}
