+++ src/components/TB/PosiList.jsx
@@ -3,8 +3,8 @@ import { useSelector, useDispatch } from 'react-redux';
 
 import styled from 'styled-components';
 
-import { newEmpId } from '../../utils/Slice';
-import { changeEmpApi } from '../../utils/API';
+import { newEmpId, newCompId } from '../../utils/Slice';
+import { changeEmpApi } from '../../api/login';
 
 export default function PosiList(props) {
 
@@ -21,7 +21,11 @@ export default function PosiList(props) {
 
   const handleAnotherEmp = () => {
     changeEmpApi(radioEmpId).then(res => {
+      console.log('chage : ', res.data['empId'], res.data['compId']);
       dispatch(newEmpId(res.data['empId']));
+      dispatch(newCompId(res.data['compId']));
+      localStorage.setItem('empId', res.data['empId']);
+      localStorage.setItem('compId', res.data['compId']);
     }).then(props.modalOff);
   }
